requirements.txt
anyascii==0.1.7
appdirs==1.4.4
asgiref==3.3.1
beautifulsoup4==4.8.2
CacheControl==0.12.6
cachy==0.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cleo==0.8.1
clikit==0.6.2
crashtest==0.3.1
cryptography==3.4.6
distlib==0.3.1
Django==3.0.13
django-bootstrap==0.2.4
django-bootstrap4==2.3.1
django-cogwheels==0.3
django-colorful==1.3
django-el-pagination==3.3.0
django-filter==2.4.0
django-modelcluster==5.1
django-social-share==2.2.1
django-taggit==1.3.0
django-treebeard==4.4
djangorestframework==3.12.2
draftjs-exporter==2.1.7
et-xmlfile==1.0.1
filelock==3.0.12
html5lib==1.1
idna==2.10
jdcal==1.4.1
jeepney==0.6.0
keyring==21.8.0
l18n==2020.6.1
lockfile==0.12.2
msgpack==1.0.2
openpyxl==3.0.6
packaging==20.9
pastel==0.2.1
pexpect==4.8.0
Pillow==8.1.1
pkg-resources==0.0.0
pkginfo==1.7.0
poetry-core==1.0.2
ptyprocess==0.7.0
puput==1.1.1
pycparser==2.20
pylev==1.3.0
pyparsing==2.4.7
python-decouple==3.4
pytz==2021.1
requests==2.25.1
requests-toolbelt==0.9.1
SecretStorage==3.3.1
shellingham==1.4.0
six==1.15.0
soupsieve==2.2
sqlparse==0.4.1
tablib==3.0.0
text-unidecode==1.3
tomlkit==0.7.0
urllib3==1.26.3
virtualenv==20.4.2
wagtail==2.12.3
wagtailmenus==3.0.2
webencodings==0.5.1
Willow==1.4
xlrd==2.0.1
XlsxWriter==1.3.7
xlwt==1.3.0
